fix(swr): fix bug in NQAQCONN counter
NQAQCONN counter was incorrect for cases where a reach was connected to
more than one layer. Changes to gwf2swt7.f and utl7.f for compiling
with gfortran. Added makefile and makedefaults to compile using GNU
make with gfortran compiler on Windows and both the gfortran and ifort
compilers on macOS and Linux. Manual adjustment of the makefile is not
required to use on different OSs or compilers. To use the makefile,
openspec.inc needs to be swapped with openspec_make.inc. Syntax for
using the makefile, after swapping the openspec.inc files, is

Windows with gfortran compiler

    make

macOS and Linux with gfortran and ifort compilers

    make -FC=gfortran -f makefile

    make -FC=ifort -f makefile